src/main/webapp/app/entities/scans/delete/scans-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { IScans } from '../scans.model';
import { ScansService } from '../service/scans.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './scans-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class ScansDeleteDialogComponent {
  scans?: IScans;

  constructor(protected scansService: ScansService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: string): void {
    this.scansService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
